Remove unused fields, fix incorrect dead_code annotations, fix clippy issues
Remove populated-but-never-consumed fields: mapq_distribution and
primary_paired from BamStat, fragment_count/transcript_coverage/
merged_gene_coverage from QualimapResult, and supporting accumulator
code. Remove unused ExonMeta.transcript_idx, MergedGeneModel.gene_idx/
exonic_length, QualimapIndex.num_genes/merged_gene_models, and
TranscriptCoverage::iter(). Fix incorrect dead_code annotations with
accurate comments. Fix clippy issues: .clone() on Copy types, unnecessary
cast, useless vec! in tests.

+++ src/rna/rseqc/idxstats.rs
@@ -73,7 +73,7 @@ mod tests {
         let header_refs: Vec<(String, u64)> = (0..header.target_count())
             .map(|tid| {
                 let name = String::from_utf8_lossy(header.tid2name(tid)).to_string();
-                let len = header.target_len(tid).unwrap_or(0) as u64;
+                let len = header.target_len(tid).unwrap_or(0);
                 (name, len)
             })
             .collect();
